backend/requirements.txt
flask
flask-cors
tensorflow==2.11.0
pillow
